Splitted share with organization component
