+++ opencti-platform/opencti-graphql/tests/03-integration/01-database/playbook-components-test.ts
@@ -1,5 +1,4 @@
 import { afterAll, beforeAll, describe, expect, it, vi } from 'vitest';
-import { PLAYBOOK_SHARING_COMPONENT } from '../../../src/modules/playbook/playbook-components';
 import type { StixBundle } from '../../../src/types/stix-2-1-common';
 import type { BasicStoreEntityOrganization } from '../../../src/modules/organization/organization-types';
 import { unSetOrganization, setOrganization } from '../../utils/testQueryHelper';
@@ -10,6 +9,7 @@ import { container_wrapper_component_bundle } from './playbookComponents/playboo
 import { STIX_EXT_OCTI } from '../../../src/types/stix-2-1-extensions';
 import * as entrepriseEdition from '../../../src/enterprise-edition/ee';
 import { PLAYBOOK_CONTAINER_WRAPPER_COMPONENT } from '../../../src/modules/playbook/components/container-wrapper-component';
+import { PLAYBOOK_SHARING_COMPONENT } from '../../../src/modules/playbook/components/share-with-organization';
 
 describe('playbook sharing component', () => {
   let externalOrganizationEntity: BasicStoreEntityOrganization;
